Feature/testing

+++ src/main/java/iss/tim4/domain/dto/working/hours/WorkingHoursDTOResult.java
@@ -1,5 +1,6 @@
 package iss.tim4.domain.dto.working.hours;
 
+import iss.tim4.domain.dto.driver.DriverDTOResult;
 import iss.tim4.domain.model.WorkingHours;
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -12,7 +13,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class WorkingHoursDTOResult {
 
-    private Long id;
+    private Integer id;
     private LocalDateTime start;
     private LocalDateTime end;
